+++ pkg/helm/client.go
@@ -5,9 +5,12 @@ import (
 	"path/filepath"
 
 	"github.com/containerd/containerd/log"
+	"groundcover.com/pkg/k8s"
 	"groundcover.com/pkg/utils"
 	"helm.sh/helm/v3/pkg/action"
 	"helm.sh/helm/v3/pkg/cli"
+	"k8s.io/cli-runtime/pkg/genericclioptions"
+	"k8s.io/client-go/rest"
 )
 
 type Client struct {
@@ -33,7 +36,14 @@ func NewHelmClient(namespace, kubecontext string) (*Client, error) {
 	helmClient.settings.SetNamespace(namespace)
 	helmClient.settings.KubeContext = kubecontext
 
-	if err = helmClient.cfg.Init(helmClient.settings.RESTClientGetter(), helmClient.settings.Namespace(), os.Getenv("HELM_DRIVER"), log.L.Debugf); err != nil {
+	getter := helmClient.settings.RESTClientGetter().(*genericclioptions.ConfigFlags)
+
+	getter.WrapConfigFn = func(restConfig *rest.Config) *rest.Config {
+		k8s.OverrideDepartedAuthenticationApiVersion(restConfig)
+		return restConfig
+	}
+
+	if err = helmClient.cfg.Init(getter, helmClient.settings.Namespace(), os.Getenv("HELM_DRIVER"), log.L.Debugf); err != nil {
 		return nil, err
 	}